+++ packages/cojson/src/tests/coValueState.test.ts
@@ -1,128 +1,114 @@
-import { beforeEach, describe, expect, onTestFinished, test, vi } from "vitest";
+import {
+  assert,
+  afterEach,
+  beforeEach,
+  describe,
+  expect,
+  onTestFinished,
+  test,
+  vi,
+} from "vitest";
 import { PeerState } from "../PeerState";
 import { CoValueCore } from "../coValueCore";
-import { CO_VALUE_LOADING_CONFIG, CoValueState } from "../coValueState";
+import { CoValueState } from "../coValueState";
 import { RawCoID } from "../ids";
 import { Peer } from "../sync";
+import { createTestMetricReader, tearDownTestMetricReader } from "./testUtils";
 
-const initialMaxRetries = CO_VALUE_LOADING_CONFIG.MAX_RETRIES;
-
-function mockMaxRetries(maxRetries: number) {
-  CO_VALUE_LOADING_CONFIG.MAX_RETRIES = maxRetries;
-
-  onTestFinished(() => {
-    CO_VALUE_LOADING_CONFIG.MAX_RETRIES = initialMaxRetries;
-  });
-}
+let metricReader: ReturnType<typeof createTestMetricReader>;
 
 beforeEach(() => {
-  mockMaxRetries(5);
+  metricReader = createTestMetricReader();
+});
+
+afterEach(() => {
+  tearDownTestMetricReader();
 });
 
 describe("CoValueState", () => {
   const mockCoValueId = "co_test123" as RawCoID;
 
-  test("should create unknown state", () => {
-    const state = CoValueState.Unknown(mockCoValueId);
+  test("should create unknown state", async () => {
+    const state = new CoValueState(mockCoValueId);
 
     expect(state.id).toBe(mockCoValueId);
-    expect(state.state.type).toBe("unknown");
+    expect(state.highLevelState).toBe("unknown");
+    expect(
+      await metricReader.getMetricValue("jazz.covalues.loaded", {
+        state: "unknown",
+      }),
+    ).toBe(1);
   });
 
-  test("should create loading state", () => {
-    const peerIds = ["peer1", "peer2"];
-    const state = CoValueState.Loading(mockCoValueId, peerIds);
+  test("should create loading state", async () => {
+    const state = new CoValueState(mockCoValueId);
+    state.loadFromPeers([
+      createMockPeerState({ id: "peer1", role: "server" }),
+      createMockPeerState({ id: "peer2", role: "server" }),
+    ]);
 
     expect(state.id).toBe(mockCoValueId);
-    expect(state.state.type).toBe("loading");
+    expect(state.highLevelState).toBe("loading");
+    expect(
+      await metricReader.getMetricValue("jazz.covalues.loaded", {
+        state: "loading",
+      }),
+    ).toBe(1);
   });
 
   test("should create available state", async () => {
     const mockCoValue = createMockCoValueCore(mockCoValueId);
-    const state = CoValueState.Available(mockCoValue);
+    const state = new CoValueState(mockCoValueId);
+    state.internalMarkMagicallyAvailable(mockCoValue);
 
     expect(state.id).toBe(mockCoValueId);
-    expect(state.state.type).toBe("available");
-    expect((state.state as any).coValue).toBe(mockCoValue);
+    expect(state.highLevelState).toBe("available");
+    expect(state.core).toBe(mockCoValue);
     await expect(state.getCoValue()).resolves.toEqual(mockCoValue);
+    expect(
+      await metricReader.getMetricValue("jazz.covalues.loaded", {
+        state: "available",
+      }),
+    ).toBe(1);
   });
 
   test("should handle found action", async () => {
     const mockCoValue = createMockCoValueCore(mockCoValueId);
-    const state = CoValueState.Loading(mockCoValueId, ["peer1", "peer2"]);
+    const state = new CoValueState(mockCoValueId);
+    state.loadFromPeers([
+      createMockPeerState({ id: "peer1", role: "server" }),
+      createMockPeerState({ id: "peer2", role: "server" }),
+    ]);
+
+    expect(
+      await metricReader.getMetricValue("jazz.covalues.loaded", {
+        state: "available",
+      }),
+    ).toBe(undefined);
+    expect(
+      await metricReader.getMetricValue("jazz.covalues.loaded", {
+        state: "loading",
+      }),
+    ).toBe(1);
 
     const stateValuePromise = state.getCoValue();
 
-    state.dispatch({
-      type: "available",
-      coValue: mockCoValue,
-    });
+    state.internalMarkMagicallyAvailable(mockCoValue);
 
     const result = await state.getCoValue();
     expect(result).toBe(mockCoValue);
     await expect(stateValuePromise).resolves.toBe(mockCoValue);
-  });
 
-  test("should ignore actions when not in loading state", () => {
-    const state = CoValueState.Unknown(mockCoValueId);
-
-    state.dispatch({
-      type: "not-found-in-peer",
-      peerId: "peer1",
-    });
-
-    expect(state.state.type).toBe("unknown");
-  });
-
-  test("should retry loading from peers when unsuccessful", async () => {
-    vi.useFakeTimers();
-
-    const peer1 = createMockPeerState(
-      {
-        id: "peer1",
-        role: "server",
-      },
-      async () => {
-        state.dispatch({
-          type: "not-found-in-peer",
-          peerId: "peer1",
-        });
-      },
-    );
-    const peer2 = createMockPeerState(
-      {
-        id: "peer2",
-        role: "server",
-      },
-      async () => {
-        state.dispatch({
-          type: "not-found-in-peer",
-          peerId: "peer2",
-        });
-      },
-    );
-    const mockPeers = [peer1, peer2] as unknown as PeerState[];
-
-    const state = CoValueState.Unknown(mockCoValueId);
-    const loadPromise = state.loadFromPeers(mockPeers);
-
-    // Should attempt CO_VALUE_LOADING_CONFIG.MAX_RETRIES retries
-    for (let i = 0; i < CO_VALUE_LOADING_CONFIG.MAX_RETRIES; i++) {
-      await vi.runAllTimersAsync();
-    }
-
-    await loadPromise;
-
-    expect(peer1.pushOutgoingMessage).toHaveBeenCalledTimes(
-      CO_VALUE_LOADING_CONFIG.MAX_RETRIES,
-    );
-    expect(peer2.pushOutgoingMessage).toHaveBeenCalledTimes(
-      CO_VALUE_LOADING_CONFIG.MAX_RETRIES,
-    );
-    expect(state.state.type).toBe("unavailable");
-    await expect(state.getCoValue()).resolves.toBe("unavailable");
-
-    vi.useRealTimers();
+    expect(
+      await metricReader.getMetricValue("jazz.covalues.loaded", {
+        state: "available",
+      }),
+    ).toBe(1);
+    expect(
+      await metricReader.getMetricValue("jazz.covalues.loaded", {
+        state: "loading",
+      }),
+    ).toBe(0);
   });
 
   test("should skip errored coValues when loading from peers", async () => {
@@ -134,11 +120,7 @@ describe("CoValueState", () => {
         role: "server",
       },
       async () => {
-        peer1.erroredCoValues.set(mockCoValueId, new Error("test") as any);
-        state.dispatch({
-          type: "not-found-in-peer",
-          peerId: "peer1",
-        });
+        state.markErrored("peer1", {} as any);
       },
     );
     const peer2 = createMockPeerState(
@@ -147,129 +129,24 @@ describe("CoValueState", () => {
         role: "server",
       },
       async () => {
-        state.dispatch({
-          type: "not-found-in-peer",
-          peerId: "peer2",
-        });
+        state.markNotFoundInPeer("peer2");
       },
     );
 
     const mockPeers = [peer1, peer2] as unknown as PeerState[];
 
-    const state = CoValueState.Unknown(mockCoValueId);
+    const state = new CoValueState(mockCoValueId);
     const loadPromise = state.loadFromPeers(mockPeers);
 
-    // Should attempt CO_VALUE_LOADING_CONFIG.MAX_RETRIES retries
-    for (let i = 0; i < CO_VALUE_LOADING_CONFIG.MAX_RETRIES; i++) {
-      await vi.runAllTimersAsync();
-    }
+    await vi.runAllTimersAsync();
 
     await loadPromise;
 
     expect(peer1.pushOutgoingMessage).toHaveBeenCalledTimes(1);
-    expect(peer2.pushOutgoingMessage).toHaveBeenCalledTimes(
-      CO_VALUE_LOADING_CONFIG.MAX_RETRIES,
-    );
-    expect(state.state.type).toBe("unavailable");
+    expect(peer2.pushOutgoingMessage).toHaveBeenCalledTimes(1);
+    expect(state.highLevelState).toBe("unavailable");
     await expect(state.getCoValue()).resolves.toBe("unavailable");
 
-    vi.useRealTimers();
-  });
-
-  test("should retry only on server peers", async () => {
-    vi.useFakeTimers();
-
-    const peer1 = createMockPeerState(
-      {
-        id: "peer1",
-        role: "storage",
-      },
-      async () => {
-        state.dispatch({
-          type: "not-found-in-peer",
-          peerId: "peer1",
-        });
-      },
-    );
-    const peer2 = createMockPeerState(
-      {
-        id: "peer2",
-        role: "server",
-      },
-      async () => {
-        state.dispatch({
-          type: "not-found-in-peer",
-          peerId: "peer2",
-        });
-      },
-    );
-    const mockPeers = [peer1, peer2] as unknown as PeerState[];
-
-    const state = CoValueState.Unknown(mockCoValueId);
-    const loadPromise = state.loadFromPeers(mockPeers);
-
-    // Should attempt CO_VALUE_LOADING_CONFIG.MAX_RETRIES retries
-    for (let i = 0; i < CO_VALUE_LOADING_CONFIG.MAX_RETRIES; i++) {
-      await vi.runAllTimersAsync();
-    }
-
-    await loadPromise;
-
-    expect(peer1.pushOutgoingMessage).toHaveBeenCalledTimes(1);
-    expect(peer2.pushOutgoingMessage).toHaveBeenCalledTimes(
-      CO_VALUE_LOADING_CONFIG.MAX_RETRIES,
-    );
-    expect(state.state.type).toBe("unavailable");
-    await expect(state.getCoValue()).resolves.toEqual("unavailable");
-
-    vi.useRealTimers();
-  });
-
-  test("should handle the coValues that become available in between of the retries", async () => {
-    vi.useFakeTimers();
-
-    mockMaxRetries(5);
-
-    let retries = 0;
-
-    const peer1 = createMockPeerState(
-      {
-        id: "peer1",
-        role: "server",
-      },
-      async () => {
-        retries++;
-        state.dispatch({
-          type: "not-found-in-peer",
-          peerId: "peer1",
-        });
-
-        if (retries === 2) {
-          setTimeout(() => {
-            state.dispatch({
-              type: "available",
-              coValue: createMockCoValueCore(mockCoValueId),
-            });
-          }, 100);
-        }
-      },
-    );
-
-    const mockPeers = [peer1] as unknown as PeerState[];
-
-    const state = CoValueState.Unknown(mockCoValueId);
-    const loadPromise = state.loadFromPeers(mockPeers);
-
-    // Should attempt CO_VALUE_LOADING_CONFIG.MAX_RETRIES retries
-    for (let i = 0; i < CO_VALUE_LOADING_CONFIG.MAX_RETRIES + 1; i++) {
-      await vi.runAllTimersAsync();
-    }
-
-    await loadPromise;
-
-    expect(peer1.pushOutgoingMessage).toHaveBeenCalledTimes(2);
-    expect(state.state.type).toBe("available");
-    await expect(state.getCoValue()).resolves.toEqual({ id: mockCoValueId });
     vi.useRealTimers();
   });
 
@@ -282,78 +159,23 @@ describe("CoValueState", () => {
         role: "server",
       },
       async () => {
-        state.dispatch({
-          type: "not-found-in-peer",
-          peerId: "peer1",
-        });
+        state.markNotFoundInPeer("peer1");
       },
     );
 
     const mockPeers = [peer1] as unknown as PeerState[];
 
-    const state = CoValueState.Unknown(mockCoValueId);
+    const state = new CoValueState(mockCoValueId);
     const loadPromise = state.loadFromPeers(mockPeers);
 
-    // Should attempt CO_VALUE_LOADING_CONFIG.MAX_RETRIES retries
-    for (let i = 0; i < CO_VALUE_LOADING_CONFIG.MAX_RETRIES; i++) {
-      await vi.runAllTimersAsync();
-    }
+    await vi.runAllTimersAsync();
 
-    state.dispatch({
-      type: "available",
-      coValue: createMockCoValueCore(mockCoValueId),
-    });
+    state.internalMarkMagicallyAvailable(createMockCoValueCore(mockCoValueId));
 
     await loadPromise;
 
-    expect(peer1.pushOutgoingMessage).toHaveBeenCalledTimes(
-      CO_VALUE_LOADING_CONFIG.MAX_RETRIES,
-    );
-    expect(state.state.type).toBe("available");
-    await expect(state.getCoValue()).resolves.toEqual({ id: mockCoValueId });
-
-    vi.useRealTimers();
-  });
-
-  test("should stop retrying when value becomes available", async () => {
-    vi.useFakeTimers();
-
-    mockMaxRetries(5);
-
-    let run = 1;
-
-    const peer1 = createMockPeerState(
-      {
-        id: "peer1",
-        role: "server",
-      },
-      async () => {
-        if (run > 2) {
-          state.dispatch({
-            type: "available",
-            coValue: createMockCoValueCore(mockCoValueId),
-          });
-        }
-        state.dispatch({
-          type: "not-found-in-peer",
-          peerId: "peer1",
-        });
-        run++;
-      },
-    );
-
-    const mockPeers = [peer1] as unknown as PeerState[];
-
-    const state = CoValueState.Unknown(mockCoValueId);
-    const loadPromise = state.loadFromPeers(mockPeers);
-
-    for (let i = 0; i < CO_VALUE_LOADING_CONFIG.MAX_RETRIES; i++) {
-      await vi.runAllTimersAsync();
-    }
-    await loadPromise;
-
-    expect(peer1.pushOutgoingMessage).toHaveBeenCalledTimes(3);
-    expect(state.state.type).toBe("available");
+    expect(peer1.pushOutgoingMessage).toHaveBeenCalledTimes(1);
+    expect(state.highLevelState).toBe("available");
     await expect(state.getCoValue()).resolves.toEqual({ id: mockCoValueId });
 
     vi.useRealTimers();
@@ -370,31 +192,23 @@ describe("CoValueState", () => {
         role: "storage",
       },
       async () => {
-        state.dispatch({
-          type: "available",
-          coValue: mockCoValue,
-        });
+        state.markAvailable(mockCoValue, "peer1");
       },
     );
     const peer2 = createMockPeerState(
       {
-        id: "peer1",
+        id: "peer2",
         role: "server",
       },
       async () => {
-        state.dispatch({
-          type: "not-found-in-peer",
-          peerId: "peer2",
-        });
+        state.markNotFoundInPeer("peer2");
       },
     );
 
-    const state = CoValueState.Unknown(mockCoValueId);
+    const state = new CoValueState(mockCoValueId);
     const loadPromise = state.loadFromPeers([peer1, peer2]);
 
-    for (let i = 0; i < CO_VALUE_LOADING_CONFIG.MAX_RETRIES; i++) {
-      await vi.runAllTimersAsync();
-    }
+    await vi.runAllTimersAsync();
     await loadPromise;
 
     expect(peer1.pushOutgoingMessage).toHaveBeenCalledTimes(1);
@@ -403,7 +217,7 @@ describe("CoValueState", () => {
       action: "load",
       ...mockCoValue.knownState(),
     });
-    expect(state.state.type).toBe("available");
+    expect(state.highLevelState).toBe("available");
     await expect(state.getCoValue()).resolves.toEqual({ id: mockCoValueId });
 
     vi.useRealTimers();
@@ -425,31 +239,26 @@ describe("CoValueState", () => {
     );
     const peer2 = createMockPeerState(
       {
-        id: "peer1",
+        id: "peer2",
         role: "server",
       },
       async () => {
-        state.dispatch({
-          type: "available",
-          coValue: mockCoValue,
-        });
+        state.markAvailable(mockCoValue, "peer2");
       },
     );
 
     peer1.closed = true;
 
-    const state = CoValueState.Unknown(mockCoValueId);
+    const state = new CoValueState(mockCoValueId);
     const loadPromise = state.loadFromPeers([peer1, peer2]);
 
-    for (let i = 0; i < CO_VALUE_LOADING_CONFIG.MAX_RETRIES; i++) {
-      await vi.runAllTimersAsync();
-    }
+    await vi.runAllTimersAsync();
     await loadPromise;
 
     expect(peer1.pushOutgoingMessage).toHaveBeenCalledTimes(0);
     expect(peer2.pushOutgoingMessage).toHaveBeenCalledTimes(1);
 
-    expect(state.state.type).toBe("available");
+    expect(state.highLevelState).toBe("available");
     await expect(state.getCoValue()).resolves.toEqual({ id: mockCoValueId });
 
     vi.useRealTimers();
@@ -466,19 +275,15 @@ describe("CoValueState", () => {
       async () => {},
     );
 
-    const state = CoValueState.Unknown(mockCoValueId);
+    const state = new CoValueState(mockCoValueId);
     const loadPromise = state.loadFromPeers([peer1]);
 
-    for (let i = 0; i < CO_VALUE_LOADING_CONFIG.MAX_RETRIES * 2; i++) {
-      await vi.runAllTimersAsync();
-    }
+    await vi.runAllTimersAsync();
     await loadPromise;
 
-    expect(peer1.pushOutgoingMessage).toHaveBeenCalledTimes(
-      CO_VALUE_LOADING_CONFIG.MAX_RETRIES,
-    );
+    expect(peer1.pushOutgoingMessage).toHaveBeenCalledTimes(1);
 
-    expect(state.state.type).toBe("unavailable");
+    expect(state.highLevelState).toBe("unavailable");
     await expect(state.getCoValue()).resolves.toEqual("unavailable");
 
     vi.useRealTimers();