+++ packages/jazz-expo/CHANGELOG.md
@@ -1,5 +1,127 @@
 # jazz-browser
 
+## 0.13.16
+
+### Patch Changes
+
+- Updated dependencies [c6fb8dc]
+  - cojson@0.13.16
+  - cojson-transport-ws@0.13.16
+  - jazz-auth-clerk@0.13.16
+  - jazz-react-core@0.13.16
+  - jazz-react-native-core@0.13.16
+  - jazz-tools@0.13.16
+
+## 0.13.15
+
+### Patch Changes
+
+- Updated dependencies [c712ef2]
+  - cojson@0.13.15
+  - cojson-transport-ws@0.13.15
+  - jazz-auth-clerk@0.13.15
+  - jazz-react-core@0.13.15
+  - jazz-react-native-core@0.13.15
+  - jazz-tools@0.13.15
+
+## 0.13.14
+
+### Patch Changes
+
+- bd94012: Fix RN docs regarding metro config
+- Updated dependencies [5c2c7d4]
+  - cojson@0.13.14
+  - cojson-transport-ws@0.13.14
+  - jazz-auth-clerk@0.13.14
+  - jazz-react-core@0.13.14
+  - jazz-react-native-core@0.13.14
+  - jazz-tools@0.13.14
+
+## 0.13.13
+
+### Patch Changes
+
+- Updated dependencies [ec9cb40]
+  - cojson@0.13.13
+  - cojson-transport-ws@0.13.13
+  - jazz-auth-clerk@0.13.13
+  - jazz-react-core@0.13.13
+  - jazz-react-native-core@0.13.13
+  - jazz-tools@0.13.13
+
+## 0.13.12
+
+### Patch Changes
+
+- Updated dependencies [4547525]
+- Updated dependencies [65719f2]
+  - jazz-tools@0.13.12
+  - cojson@0.13.12
+  - jazz-auth-clerk@0.13.12
+  - jazz-react-core@0.13.12
+  - jazz-react-native-core@0.13.12
+  - cojson-transport-ws@0.13.12
+
+## 0.13.11
+
+### Patch Changes
+
+- Updated dependencies [17273a6]
+- Updated dependencies [3396ed4]
+- Updated dependencies [17273a6]
+- Updated dependencies [267ea4c]
+  - cojson@0.13.11
+  - jazz-tools@0.13.11
+  - cojson-transport-ws@0.13.11
+  - jazz-auth-clerk@0.13.11
+  - jazz-react-core@0.13.11
+  - jazz-react-native-core@0.13.11
+
+## 0.13.10
+
+### Patch Changes
+
+- Updated dependencies [f837cfe]
+  - cojson@0.13.10
+  - cojson-transport-ws@0.13.10
+  - jazz-auth-clerk@0.13.10
+  - jazz-react-core@0.13.10
+  - jazz-react-native-core@0.13.10
+  - jazz-tools@0.13.10
+
+## 0.13.9
+
+### Patch Changes
+
+- Updated dependencies [a6cf01f]
+  - jazz-tools@0.13.9
+  - jazz-auth-clerk@0.13.9
+  - jazz-react-core@0.13.9
+  - jazz-react-native-core@0.13.9
+
+## 0.13.8
+
+### Patch Changes
+
+- Updated dependencies [de19a6d]
+  - jazz-auth-clerk@0.13.8
+
+## 0.13.7
+
+### Patch Changes
+
+- Updated dependencies [bc3d7bb]
+- Updated dependencies [4e9aae1]
+- Updated dependencies [21c935c]
+- Updated dependencies [aa1c80e]
+- Updated dependencies [13074be]
+  - jazz-tools@0.13.7
+  - cojson@0.13.7
+  - jazz-auth-clerk@0.13.7
+  - jazz-react-core@0.13.7
+  - jazz-react-native-core@0.13.7
+  - cojson-transport-ws@0.13.7
+
 ## 0.13.5
 
 ### Patch Changes